Version 2.23.17
Changelog for i40e-2.23.17
===========================================================================
Add PHY debug register dump
Add mdd-auto-reset-vf private flag
Fix FW v8.15 double vlan issue
Fix 32bit FW gtime wrapping issue for nvmupdate
Fix err -110 after DCB ETS configuration
Fix pointer dereference after free
Fix i40e_setup_misc_vector() error handling
Fix missing rtnl locking when setting up pf switch
Fix accessing vsi->active_filters without holding lock
Fix SW DCB configuration when ADQ is set
Fix PTP pins verification
Fix passing tainted expression to kzalloc()
Fix error by one for traffic class number
Fix compilation issue with gcc older than 5.0.0
Fix compilation under RHEL 9.2 & 8.8

+++ src/i40e_type.h
@@ -1,5 +1,5 @@
-/* SPDX-License-Identifier: GPL-2.0 */
-/* Copyright(c) 2013 - 2023 Intel Corporation. */
+/* SPDX-License-Identifier: GPL-2.0-only */
+/* Copyright (C) 2013-2023 Intel Corporation */
 
 #ifndef _I40E_TYPE_H_
 #define _I40E_TYPE_H_
@@ -25,7 +25,6 @@
 
 /* Max timeout in ms for the phy to respond */
 #define I40E_MAX_PHY_TIMEOUT		500
-
 /* Switch from ms to the 1usec global time (this is the GTIME resolution) */
 #define I40E_MS_TO_GTIME(time)		((time) * 1000)
 
@@ -380,7 +379,7 @@ enum i40e_aq_resource_access_type {
 };
 
 struct i40e_nvm_info {
-	u64 hw_semaphore_timeout; /* usec global time (GTIME resolution) */
+	u32 hw_semaphore_timeout; /* usec global time (GTIME resolution) */
 	u32 timeout;              /* [ms] */
 	u16 sr_size;              /* Shadow RAM size in words */
 	bool blank_nvm_mode;      /* is NVM empty (no FW present)*/
